backend/requirements.txt
amqp==5.1.1
apturl==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
billiard==3.6.4.0
black==22.12.0
blinker==1.4
Brlapi==0.7.0
celery==5.2.7
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
dnspython==2.2.1
duplicity==0.8.12.0
ecdsa==0.18.0
entrypoints==0.3
enum-compat==0.0.3
eventlet==0.33.2
fasteners==0.14.1
filelock==3.9.0
future==0.18.2
greenlet==2.0.1
httplib2==0.14.0
idna==2.8
intuit-oauth==1.2.4
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
mypy-extensions==0.4.3
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pathspec==0.10.3
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==3.6.1
pyasn1==0.4.8
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-jose==3.3.0
pytz==2022.6
pyxdg==0.26
PyYAML==5.3.1
redis==4.4.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.3
systemd-python==234
tomli==2.0.1
typing-extensions==4.4.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.17.1
wadllib==1.3.3
wcwidth==0.2.5
xkit==0.0.0
